+++ demo2/src/main/java/Demo1Application.java
@@ -1,0 +1,5 @@
+//import org.springframework.boot.autoconfigure.SpringBootApplication;
+//
+//@SpringBootApplication
+//public class Demo1Application {
+//}
