--- demo2/src/main/java/service/SampleService.java
@@ -1,4 +1,0 @@
-package service;
-
-public class SampleService {
-}
