+++ demo2/src/main/resources/application.properties
@@ -1,9 +1,11 @@
 spring.application.name=demo2
 spring.datasource.driver-class-name=org.mariadb.jdbc.Driver
-spring.datasource.url=jdbc:mariadb://3.35.197.144:3306
+spring.datasource.url=jdbc:mariadb://3.35.197.144:3306/oms
 spring.datasource.username=root
 spring.datasource.password=1234
 spring.jpa.show-sql=true
 spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.MariaDBDialect
 spring.jpa.hibernate.ddl-auto=update
+logging.level.com.zaxxer.hikari=DEBUG
+
 server.port=8080
